refactor: split dones core
